# api/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "WIGPI API",
      "version": "1.0.0"
    },
    "schemes": [
      "https"
    ],
    "host": "wigpi.flusin.fr",
    "basePath": "/api",
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
...
      "/system/refresh_edt": {
        "get": {
          "summary": "GET /system/refresh_edt",
          "operationId": "RefreshEDT",
          "parameters": [
            {
              "name": "apiKey",
              "in": "query",
              "required": true,
              "type": "string",
              "format": "string",
              "description": "API key to authenticate the request"
            }
          ],
          "responses": {
            "200": {
              "description": "EDT refresh triggered successfully",
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "EDT refresh triggered successfully"
                  }
                }
              }
            },
            "403": {
              "description": "Invalid or inactive API key",
              "schema": {
                "type": "object",
                "properties": {
                  "error": {
                    "type": "string",
                    "example": "Invalid or inactive API key"
                  }
                }
              }
            },
            "409": {
              "description": "EDT refresh already in progress",
              "schema": {
                "type": "object",
                "properties": {
                  "error": {
                    "type": "string",
                    "example": "EDT refresh already in progress"
                  }
                }
              }
            },
            "500": {
              "description": "Failed to log request",
              "schema": {
                "type": "object",
                "properties": {
                  "error": {
                    "type": "string",
                    "example": "Failed to log request"
                  }
                }
              }
            }
          },
          "tags": [
            "System"
          ]
        }
      },
...
    },
...
  }